uppdaterad med fungerande tillstånd
