+++ lib/NewTodoView.dart
@@ -1,0 +1,124 @@
+import 'package:flutter/material.dart';
+import 'package:reminder_app/model.dart';
+
+class NewTodoView extends StatefulWidget {
+  final TodoCard card;
+
+  NewTodoView(this.card);
+
+  @override
+  State<StatefulWidget> createState() {
+    return NewTodoViewState(card);
+  }
+}
+
+class NewTodoViewState extends State<NewTodoView> {
+  String message;
+
+  TextEditingController textEditingController;
+
+  NewTodoViewState(TodoCard card) {
+    this.message = card.message;
+
+    textEditingController = TextEditingController(text: card.message);
+
+    textEditingController.addListener(() {
+      setState(() {
+        message = textEditingController.text;
+      });
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.black,
+      body: SafeArea(
+        child: Column(
+          children: <Widget>[_topheader(context), _searchbox(context)],
+        ),
+      ),
+    );
+  }
+
+  Widget _topheader(BuildContext context) {
+    return Column(
+      children: <Widget>[
+        Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: <Widget>[
+              Text(
+                'Lägg till ny påminnelse',
+                style: Theme.of(context)
+                    .textTheme
+                    .headline5
+                    .copyWith(fontWeight: FontWeight.bold),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _searchbox(BuildContext context) {
+    return Column(
+      children: <Widget>[
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 8.0),
+          child: Row(
+            children: <Widget>[
+              Expanded(
+                child: Card(
+                  child: Padding(
+                    padding: const EdgeInsets.all(12.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: <Widget>[
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: <Widget>[
+                            Expanded(
+                              child: TextField(
+                                controller: textEditingController,
+                                decoration: InputDecoration(
+                                  enabledBorder: InputBorder.none,
+                                  focusedBorder: InputBorder.none,
+                                  hintText: 'Skriv här...',
+                                  border: OutlineInputBorder(),
+                                ),
+                              ),
+                            ),
+                            Padding(
+                              padding: const EdgeInsets.only(left: 10.0),
+                              child: IconButton(
+                                icon: Icon(Icons.add),
+                                onPressed: () {
+                                  Navigator.pop(
+                                    context,
+                                    TodoCard(
+                                      message: message,
+                                    ),
+                                  );
+                                },
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
+                  ),
+                  color: Colors.white10,
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(10.0)),
+                ),
+              ),
+            ],
+          ),
+        )
+      ],
+    );
+  }
+}
